Change the run script
The run script now accepts a single argument of which properties file to
use. Also it is now much clearer that the properties file is only an
example, so you'll need to supply your own with your real values to use.
To facilitate creating your own properties file compatible with your
container registry, we are now ignoring files named local.properties.