Delete all images
With a small change to the delete script, we can delete all images no
matter the tag. This has how been integrated into a global cleanup step
at the end of the performance test.

This is needed because while it does help to remove all the individual
tags, there are still some remaining artifacts that need purging at the
end of a test run.